+++ src/zepben/evolve/database/sqlite/tables/iec61970/base/core/table_identified_objects.py
@@ -5,7 +5,7 @@
 from abc import ABC
 from typing import List, Generator
 
-from zepben.evolve.database.sqlite.tables.column import Column, Nullable
+from zepben.evolve.database.sql.column import Column, Nullable
 from zepben.evolve.database.sqlite.tables.sqlite_table import SqliteTable