+++ src/zepben/evolve/database/sqlite/tables/iec61970/base/wires/table_regulating_cond_eq.py
@@ -4,7 +4,7 @@
 #  file, You can obtain one at https://mozilla.org/MPL/2.0/.
 from abc import ABC
 
-from zepben.evolve.database.sqlite.tables.column import Column, Nullable
+from zepben.evolve.database.sql.column import Column, Nullable
 from zepben.evolve.database.sqlite.tables.iec61970.base.wires.table_energy_connections import TableEnergyConnections
 
 __all__ = ["TableRegulatingCondEq"]